requirements.txt
requests
openai
python-dotenv
fastapi
uvicorn
python-multipart
aiofiles
webrtcvad
webrtcvad
